+++ src/imio/annex/browser/views.py
@@ -2,20 +2,66 @@
 
 from collective.eeafaceted.batchactions import _ as _CEBA
 from collective.eeafaceted.batchactions.browser.views import BaseBatchActionForm
-from imio.annex import _
+from collective.iconifiedcategory.utils import calculate_filesize
 from imio.annex.content.annex import IAnnex
 from io import BytesIO
 from plone import api
 from plone.rfc822.interfaces import IPrimaryFieldInfo
+from zope.i18n import translate
 
 import zipfile
+
+MAX_TOTAL_SIZE = 50000000
 
 
 class DownloadAnnexesBatchActionForm(BaseBatchActionForm):
 
     label = _CEBA("Download annexes")
-    button_with_icon = False
+    button_with_icon = True
     section = "annexes"
+
+    @property
+    def description(self):
+        """ """
+        descr = super(DownloadAnnexesBatchActionForm, self).description
+        descr = translate(descr, domain=descr.domain, context=self.request)
+        readable_total_size = calculate_filesize(self.total_size)
+        readable_max_size = calculate_filesize(MAX_TOTAL_SIZE)
+        if self.total_size > MAX_TOTAL_SIZE:
+            descr += translate(
+                '<p class="warn_filesize">The maximum size you may download at one '
+                'time is ${max_size}, here your download size is ${total_size}. '
+                'Please unselect some elements, especially large elements for which '
+                'size is displayed in red, download it separately.<p>',
+                mapping={'max_size': readable_max_size,
+                         'total_size': readable_total_size},
+                domain="collective.eeafaceted.batchactions",
+                context=self.request)
+            self.do_apply = False
+        else:
+            descr += translate(
+                '<p>This will download the selected elements as a Zip file.</p>'
+                '<p>The total file size is <b>${total_size}</b>, when clicking on "Apply", '
+                'you will have a spinner, wait until the Zip file is available.</p>',
+                mapping={'total_size': readable_total_size},
+                domain="collective.eeafaceted.batchactions",
+                context=self.request)
+        return descr
+
+    def _total_size(self):
+        """ """
+        total = 0
+        for brain in self.brains:
+            obj = brain.getObject()
+            primary_field = IPrimaryFieldInfo(obj)
+            size = primary_field.value.size
+            total += size
+        return total
+
+    def _update(self):
+        self.total_size = self._total_size()
+        if self.total_size > MAX_TOTAL_SIZE:
+            self.do_apply = False
 
     def available(self):
         """ """
@@ -44,7 +90,7 @@ class DownloadAnnexesBatchActionForm(BaseBatchActionForm):
         try:
             return self.do_zip()
         except zipfile.LargeZipFile:
-            message = _("Too much annexes to zip, try selecting fewer annexes...")
+            message = "Too much annexes to zip, try selecting fewer annexes..."
             api.portal.show_message(message, self.request, type="error")
             return self.request.response.redirect(self.context.absolute_url())