+++ CHANGES.rst
@@ -10,7 +10,9 @@ Changelog
   column so we have CSS classes to skin the column.
   [gbastien]
 - Added `DownloadAnnexesBatchActionForm`, a batch action to download several
-  annexes as a Zip file.
+  annexes as a Zip file :
+  - download is handled by an ajax request;
+  - max download size is 50Mb by default.
   [gbastien]
 
 2.8 (2021-04-23)